fix: props now are updated
