+++ libs/frontend/modules/builder/src/renderer/utils/onRendered.ts
@@ -1,6 +1,40 @@
 import { IElement } from '@codelab/shared/abstract/core'
-import { ReactNode } from 'react'
+import { ReactElement, ReactNode } from 'react'
 import { RenderContext, RenderOutput } from '../types/RenderTypes'
+
+export const updateLastRenderProps = (
+  context: RenderContext,
+  rendered: RenderOutput,
+) => {
+  if (context.onRendered) {
+    const renderMap: Record<string, ReactElement> = {}
+    const queue = [rendered]
+
+    while (queue.length) {
+      const element = queue.shift()
+
+      if (element) {
+        if (Array.isArray(element)) {
+          queue.push(...element)
+          continue
+        }
+
+        const props = (element as ReactElement)?.props
+        const id = props?.['data-id']
+
+        if (id) {
+          renderMap[id] = element as ReactElement
+        }
+
+        if (props.children) {
+          queue.push(props.children)
+        }
+      }
+    }
+
+    context.onRendered(renderMap)
+  }
+}
 
 export const onRendered = (
   rendered: RenderOutput,
@@ -8,6 +42,8 @@ export const onRendered = (
   context: RenderContext,
 ) => {
   const callRendered = (r: ReactNode) => {
+    updateLastRenderProps(context, rendered)
+
     if (context.inspect) {
       console.dir({ element: { ...element }, rendered: r })
     }
